Bump version to 2.6.0
While SPM support was added back in #32, it did not include a definition
for the YMOverrideTestSupport library. This is pretty useful for those
who like to add tests around the state of feature flags.

@cshaines updated the SPM manifest in #40, but we never published a new
release. Bump the version to 2.6.0, then we can tag and publish to both
CocoaPods and SPM.